refactor: use Q_ENUM reflection for ShortcutAction name logging
1. Move ShortcutAction enum into ShortcutController class with Q_ENUM
2. Add ShortcutController::actionName() using QMetaEnum::valueToKey
3. Add gesture registration logging to registerSwipeGesture and registerHoldGesture
4. Replace action int with actionName(action) in registerKey override log

Log: Show readable action names like "OpenMultiTaskView" in logs instead of raw integer 16

Influence:
1. Trigger gesture registration and verify log shows action name not integer
2. Register conflicting key shortcut and verify override log shows action name
3. Verify existing code with bare ShortcutAction references compiles in shortcutrunner.cpp

refactor: 使用 Q_ENUM 反射记录 ShortcutAction 名称

1. 将 ShortcutAction 枚举移入 ShortcutController 类并标记 Q_ENUM
2. 新增 ShortcutController::actionName() 通过 QMetaEnum::valueToKey 反射获取动作名
3. 为 registerSwipeGesture 和 registerHoldGesture 增加手势注册日志
4. registerKey 覆盖日志中将 action 整数替换为 actionName(action)

Log: 日志中显示可读的 action 名称如 "OpenMultiTaskView" 而非原始整数 16

Influence:
1. 触发手势注册验证日志中 action 显示为名称而非整数
2. 注册冲突快捷键验证覆盖日志中 action 显示为名称
3. 验证 shortcutrunner.cpp 中使用裸 ShortcutAction 引用的代码能正常编译

PMS: TASK-390751
